Globalized the Connection interface
