+++ app/src/hooks/use-connections.ts
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { useMatch } from "react-router-dom";
-import { ConnectionContext, Connection } from "../providers/ConnectionProvider";
+import { ConnectionContext } from "../providers/ConnectionProvider";
 
 export interface ConnectionItem {
   name: string;
