+++ app/src/providers/ConnectionProvider/index.ts
@@ -1,3 +1,1 @@
 export { ConnectionProvider, ConnectionContext } from "./ConnectionProvider";
-
-export type { Connection } from "./ConnectionProvider";
