+++ app/src/views/SchemaView/hooks/use-tables.ts
@@ -1,6 +1,4 @@
 import { useQuery } from "../../../hooks/use-query";
-import { Connection } from "../../../providers/ConnectionProvider";
-export type { Connection } from "../../../providers/ConnectionProvider";
 
 const SQL_QUERY = `
 WITH table_info AS (
